✨ feat:  page components and add about page
This commit refactors the page components by separating the home and about pages into their own files. The code changes include moving the code from the main page component to the respective page components. This improves code organization and maintainability. Additionally, the about page is added with its own content and styling.
